[Refactor](table function) Decouple vectorized table functions from non-vectorized ones
