+++ be/src/exprs/CMakeLists.txt
@@ -74,9 +74,4 @@ add_library(Exprs
   quantile_function.cpp
   grouping_sets_functions.cpp
   topn_function.cpp
-  table_function/explode_split.cpp
-  table_function/explode_bitmap.cpp
-  table_function/explode_json_array.cpp
-  table_function/table_function_factory.cpp
-  table_function/dummy_table_functions.cpp
 )
