+++ be/src/vec/exprs/table_function/vexplode_json_array.cpp
@@ -22,8 +22,108 @@
 
 namespace doris::vectorized {
 
+std::string ParsedData::true_value = "true";
+std::string ParsedData::false_value = "false";
+
+int ParsedData::set_output(ExplodeJsonArrayType type, rapidjson::Document& document) {
+    int size = document.GetArray().Size();
+    switch (type) {
+    case ExplodeJsonArrayType::INT: {
+        _data.resize(size);
+        _backup_int.resize(size);
+        int i = 0;
+        for (auto& v : document.GetArray()) {
+            if (v.IsInt64()) {
+                _backup_int[i] = v.GetInt64();
+                _data[i] = &_backup_int[i];
+            } else {
+                _data[i] = nullptr;
+            }
+            ++i;
+        }
+        break;
+    }
+    case ExplodeJsonArrayType::DOUBLE: {
+        _data.resize(size);
+        _backup_double.resize(size);
+        int i = 0;
+        for (auto& v : document.GetArray()) {
+            if (v.IsDouble()) {
+                _backup_double[i] = v.GetDouble();
+                _data[i] = &_backup_double[i];
+            } else {
+                _data[i] = nullptr;
+            }
+            ++i;
+        }
+        break;
+    }
+    case ExplodeJsonArrayType::STRING: {
+        _data_string.clear();
+        _backup_string.clear();
+        _string_nulls.clear();
+        int32_t wbytes = 0;
+        for (auto& v : document.GetArray()) {
+            switch (v.GetType()) {
+            case rapidjson::Type::kStringType:
+                _backup_string.emplace_back(v.GetString(), v.GetStringLength());
+                _string_nulls.push_back(false);
+                // do not set _data_string here.
+                // Because the address of the string stored in `_backup_string` may
+                // change each time `emplace_back()` is called.
+                break;
+            case rapidjson::Type::kNumberType:
+                if (v.IsUint()) {
+                    wbytes = snprintf(tmp_buf, sizeof(tmp_buf), "%u", v.GetUint());
+                } else if (v.IsInt()) {
+                    wbytes = snprintf(tmp_buf, sizeof(tmp_buf), "%d", v.GetInt());
+                } else if (v.IsUint64()) {
+                    wbytes = snprintf(tmp_buf, sizeof(tmp_buf), "%" PRIu64, v.GetUint64());
+                } else if (v.IsInt64()) {
+                    wbytes = snprintf(tmp_buf, sizeof(tmp_buf), "%" PRId64, v.GetInt64());
+                } else {
+                    wbytes = snprintf(tmp_buf, sizeof(tmp_buf), "%f", v.GetDouble());
+                }
+                _backup_string.emplace_back(tmp_buf, wbytes);
+                _string_nulls.push_back(false);
+                // do not set _data_string here.
+                // Because the address of the string stored in `_backup_string` may
+                // change each time `emplace_back()` is called.
+                break;
+            case rapidjson::Type::kFalseType:
+                _data_string.emplace_back(true_value);
+                _string_nulls.push_back(false);
+                break;
+            case rapidjson::Type::kTrueType:
+                _data_string.emplace_back(false_value);
+                _string_nulls.push_back(false);
+                break;
+            case rapidjson::Type::kNullType:
+                _data_string.push_back({});
+                _string_nulls.push_back(true);
+                break;
+            default:
+                _data_string.push_back({});
+                _string_nulls.push_back(true);
+                break;
+            }
+        }
+        // Must set _data_string at the end, so that we can
+        // save the real addr of string in `_backup_string` to `_data_string`.
+        for (auto& str : _backup_string) {
+            _data_string.emplace_back(str);
+        }
+        break;
+    }
+    default:
+        CHECK(false) << type;
+        break;
+    }
+    return size;
+}
+
 VExplodeJsonArrayTableFunction::VExplodeJsonArrayTableFunction(ExplodeJsonArrayType type)
-        : ExplodeJsonArrayTableFunction(type) {
+        : _type(type) {
     _fn_name = "vexplode_json_array";
 }
 
@@ -36,6 +136,12 @@ Status VExplodeJsonArrayTableFunction::process_init(vectorized::Block* block) {
                                                                    &text_column_idx));
     _text_column = block->get_by_position(text_column_idx).column;
 
+    return Status::OK();
+}
+
+Status VExplodeJsonArrayTableFunction::reset() {
+    _eos = false;
+    _cur_offset = 0;
     return Status::OK();
 }
